requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-model-utils==4.2.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
pillow==8.0.1
